# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quetzal API
  version: "1.0"
  license:
    name: MIT
  description: |
    HTTP API for Quetzal, a Kubernetes-native game-server control plane.

    ## Authentication
    Most endpoints require authentication. Two credential types are accepted:

    - **API key** (recommended for scripts): send `Authorization: Bearer <token>`.
      Create a key under *Account → API keys*; it inherits the owner's
      permissions. Tokens are shown once at creation.
    - **Session cookie**: set by `POST /api/login` and used by the web UI.

    A key/session identifies a user; per-server access follows that user's
    ownership and granted subuser permissions. Admin-only endpoints are noted.

    ## Conventions
    - Request and response bodies are JSON unless noted.
    - Mutating requests made from a browser are subject to a same-origin check.
    - Authentication endpoints are rate limited (HTTP 429 with `Retry-After`).
servers:
  - url: /
    description: Same origin as the panel
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/setup:
    post:
      tags: [Auth]
      summary: Create the first admin account (first run only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "201":
          description: Admin created (session cookie set)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "409": { description: Setup already completed }
  /api/login:
    post:
      tags: [Auth]
      summary: Log in
      description: >
        On success a session cookie is set and the user is returned. If the
        account has 2FA enabled and no `code` was supplied, the response is
        `{ "twoFactorRequired": true }` with no session; resubmit with a TOTP or
        recovery code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Authenticated, or a 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/User"
                  - type: object
                    properties:
                      twoFactorRequired: { type: boolean, enum: [true] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/me:
    get:
      tags: [Auth]
      summary: The current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/me/password:
    post:
      tags: [Users]
      summary: Change your own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string, format: password }
                newPassword: { type: string, format: password }
      responses:
        "204": { description: Password changed }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/me/email:
    put:
      tags: [Email]
      summary: Set or clear your own email (used for password reset; not verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/forgot-password:
    post:
      tags: [Email]
      summary: Request a password-reset email
      description: >-
        Always responds 200 regardless of whether the identifier matched (no
        account enumeration); the email is sent only if the account exists, has
        an email, and SMTP + the public URL are configured. Rate-limited.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { identifier: { type: string, description: username or email } }
      responses:
        "200": { description: Always returned (uniform response) }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/reset-password:
    post:
      tags: [Email]
      summary: Complete a password reset with an emailed token
      description: >-
        Validates the single-use token, sets the new password, then invalidates
        the token and all of that user's sessions. 2FA is left untouched.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, format: password, minLength: 8 }
      responses:
        "204": { description: Password reset }
        "400": { description: Invalid/expired token or weak password }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/email-settings:
    get:
      tags: [Email]
      summary: Get system SMTP settings (admin)
      description: The password is never returned, only whether one is set.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured: { type: boolean }
                  host: { type: string }
                  port: { type: string }
                  username: { type: string }
                  from: { type: string }
                  tls: { type: string }
                  hasPassword: { type: boolean }
                  publicUrl: { type: string }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [Email]
      summary: Update system SMTP settings + public URL (admin)
      description: An empty host clears the SMTP config; a blank password keeps the stored one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host: { type: string }
                port: { type: string }
                username: { type: string }
                password: { type: string, format: password }
                from: { type: string }
                tls: { type: string, enum: [starttls, tls, none] }
                publicUrl: { type: string }
      responses:
        "204": { description: Saved }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/email-settings/test:
    post:
      tags: [Email]
      summary: Send a test email with the stored settings (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { to: { type: string, description: defaults to your account email } }
      responses:
        "204": { description: Sent }
        "400": { description: Not configured or no recipient }
        "502": { description: Send failed }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/database-hosts:
    get:
      tags: [Databases]
      summary: List database hosts (admin)
      responses:
        "200": { description: OK }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Databases]
      summary: Register a database host (admin)
      description: >-
        kind=external registers an existing MySQL/MariaDB (host/port/adminUser/
        adminPassword required). kind=managed has Quetzal deploy and own a MariaDB
        in-cluster (image/storageSize/namespace optional; root password generated).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DatabaseHostInput" }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/database-hosts/{hid}:
    parameters:
      - { name: hid, in: path, required: true, schema: { type: integer } }
    patch:
      tags: [Databases]
      summary: Update a database host (admin)
      responses:
        "200": { description: Updated }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Databases]
      summary: Delete a database host (admin)
      description: Rejected with 409 while the host still holds databases.
      responses:
        "204": { description: Deleted }
        "409": { description: Host still has databases }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/database-hosts/{hid}/test:
    parameters:
      - { name: hid, in: path, required: true, schema: { type: integer } }
    post:
      tags: [Databases]
      summary: Probe a host's connectivity (admin)
      responses:
        "200": { description: Host with refreshed reachability }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/servers/{id}/database-hosts:
    parameters:
      - $ref: "#/components/parameters/ServerId"
    get:
      tags: [Databases]
      summary: Hosts a server may create databases on (minimal view)
      description: Requires the databases permission; no admin secrets returned.
      responses:
        "200": { description: OK }
  /api/servers/{id}/databases:
    parameters:
      - $ref: "#/components/parameters/ServerId"
    get:
      tags: [Databases]
      summary: List a server's databases
      description: Requires the databases permission. Passwords are not included.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ServerDatabase" } }
    post:
      tags: [Databases]
      summary: Provision a database + scoped user for the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostId]
              properties: { hostId: { type: integer } }
      responses:
        "201":
          description: Created (response includes the generated password once)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerDatabase" }
        "409": { description: Host at capacity }
        "502": { description: Provisioning failed on the host }
  /api/servers/{id}/databases/{dbid}:
    parameters:
      - $ref: "#/components/parameters/ServerId"
      - { name: dbid, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Databases]
      summary: A database with its current credentials (incl. password)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerDatabase" }
    delete:
      tags: [Databases]
      summary: Drop the database and its user
      responses:
        "204": { description: Deleted }
  /api/servers/{id}/databases/{dbid}/rotate:
    parameters:
      - $ref: "#/components/parameters/ServerId"
      - { name: dbid, in: path, required: true, schema: { type: integer } }
    post:
      tags: [Databases]
      summary: Rotate the database user's password
      responses:
        "200":
          description: New credentials (includes the new password)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerDatabase" }
  /api/me/sshkeys:
    get:
      tags: [SFTP]
      summary: List your SSH public keys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SSHKey" }
    post:
      tags: [SFTP]
      summary: Add an SSH public key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publicKey]
              properties:
                name: { type: string }
                publicKey: { type: string, description: An authorized_keys line }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SSHKey" }
        "400": { description: Invalid SSH public key }
  /api/me/sshkeys/{kid}:
    parameters:
      - { name: kid, in: path, required: true, schema: { type: integer } }
    delete:
      tags: [SFTP]
      summary: Delete one of your SSH keys
      responses:
        "204": { description: Deleted }
  /api/servers/{id}/sftp:
    parameters:
      - $ref: "#/components/parameters/ServerId"
    get:
      tags: [SFTP]
      summary: SFTP connection details for a server
      description: >-
        Requires the files permission. Port is the assigned NodePort (0 until
        Kubernetes assigns it, or if no system image is configured — the server
        status reports the latter). Auth is key-only (register keys via
        /api/me/sshkeys); the username is cosmetic. Access is confined to the
        data directory. Revoking a key cuts live sessions, but the change takes
        up to a controller resync plus ConfigMap refresh to reach the pod.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  username: { type: string }
                  port: { type: integer }
  /api/me/2fa/setup:
    post:
      tags: [TwoFactor]
      summary: Begin TOTP enrollment
      description: Returns a secret and otpauth URI; not active until enabled.
      responses:
        "200":
          description: Pending enrollment
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string }
                  uri: { type: string, description: otpauth:// provisioning URI }
        "409": { description: 2FA already enabled }
  /api/me/2fa/enable:
    post:
      tags: [TwoFactor]
      summary: Confirm enrollment with a code
      description: Activates 2FA and returns one-time recovery codes (shown once).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties: { code: { type: string } }
      responses:
        "200":
          description: 2FA enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  recoveryCodes:
                    type: array
                    items: { type: string }
        "400": { description: Invalid code or no enrollment in progress }
  /api/me/2fa/disable:
    post:
      tags: [TwoFactor]
      summary: Disable your 2FA (re-verifies a code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties: { code: { type: string } }
      responses:
        "204": { description: 2FA disabled }
        "400": { description: Invalid code or 2FA not enabled }
...
  /api/templates/{slug}/export:
    parameters:
      - { name: slug, in: path, required: true, schema: { type: string } }
    get:
      tags: [Servers]
      summary: Export a template as native JSON (admin)
      responses:
        "200": { description: Native template JSON (re-importable via PUT) }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key created under Account → API keys.
    cookieAuth:
      type: apiKey
      in: cookie
      name: quetzal_session
      description: Session cookie set by POST /api/login (used by the web UI).
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...